+++ components/vector/vector.hpp
@@ -166,7 +166,7 @@ namespace components::vector {
         // Set a nested element null/valid by path. The path is {row, sub, sub, ...}: the first
         // index is the row in this vector, each following index descends one level (list/array
         // element, or struct field). A single-element path is equivalent to set_null(row, value).
-        void set_null(const std::pmr::vector<uint64_t>& path, bool value);
+        void set_null(const std::pmr::vector<size_t>& path, bool value);
 
         void append(const vector_t& source, uint64_t source_size, uint64_t source_offset = 0);
         void append(const vector_t& source,
@@ -187,7 +187,7 @@ namespace components::vector {
         const indexing_vector_t& indexing() const;
         size_t size() const;
         bool is_null(uint64_t index = 0) const;
-        bool is_null(const std::pmr::vector<uint64_t>& path) const;
+        bool is_null(const std::pmr::vector<size_t>& path) const;
 
         void get_sequence(int64_t& start, int64_t& increment, int64_t& sequence_count) const;
         void get_sequence(int64_t& start, int64_t& increment) const;
@@ -213,7 +213,7 @@ namespace components::vector {
         // (array<array>, list<array>, ...) resolve. An out-of-range list subscript, or a NULL enclosing
         // container, is a NULL element.
         nested_element_t
-        resolve_nested_element(uint64_t row_index, const std::pmr::vector<uint64_t>& path, size_t element_start) const;
+        resolve_nested_element(uint64_t row_index, const std::pmr::vector<size_t>& path, size_t element_start) const;
 
     private:
         const vector_t* resolve_value_location(uint64_t row_index, uint64_t* index) const;
